requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0
requests>=2.28.0
python-dotenv>=0.19.0
PyYAML>=6.0
beautifulsoup4>=4.11.0
email-validator>=1.3.0